+++ src/main/java/baseball/domain/GameResult.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 import baseball.constant.GameConstant;
 
@@ -13,8 +13,8 @@ public record GameResult(int ballCnt, int strikeCnt) {
     }
 
     public boolean isBallStrike() {
-        return this.ballCnt > 0 && strikeCnt > 0 &&
-                (this.ballCnt != GameConstant.MAXIMUM_STRIKE_COUNT.getConstant() && this.strikeCnt != (GameConstant.MAXIMUM_STRIKE_COUNT.getConstant()));
+        return this.ballCnt > 0 && strikeCnt > 0 && (this.ballCnt != GameConstant.MAXIMUM_STRIKE_COUNT.getConstant()
+                        && this.strikeCnt != (GameConstant.MAXIMUM_STRIKE_COUNT.getConstant()));
     }
 
     public boolean isThreeStrike() {
