+++ src/main/java/baseball/domain/GameSet.java
@@ -1,5 +1,6 @@
-package baseball;
+package baseball.domain;
 
+import baseball.controller.GameController;
 import baseball.constant.GameConstant;
 import baseball.constant.GameMessage;
 import baseball.constant.Status;
@@ -34,11 +35,9 @@ public class GameSet {
                 GameConstant.REPLAY.getConstant(), GameConstant.QUIT.getConstant());
         int replayCommand = player.inputReplay();
         if (replayCommand == GameConstant.REPLAY.getConstant()) {
-            System.out.println("게임 리플레이");
             GameController.gameStatus = Status.GAME;
         }
         if (replayCommand == GameConstant.QUIT.getConstant()) {
-            System.out.println("게임 종료");
             GameController.gameStatus = Status.EXIT;
         }
     }
