+++ src/main/java/baseball/domain/Judge.java
@@ -1,5 +1,6 @@
-package baseball;
+package baseball.domain;
 
+import baseball.controller.GameController;
 import baseball.constant.GameConstant;
 import baseball.constant.GameMessage;
 import baseball.constant.Status;
