+++ src/main/java/baseball/domain/Player.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 import baseball.util.InputUtil;
